+++ source/config/profileUpgradeSteps.py
@@ -584,18 +584,19 @@ def upgradeConfigFrom_17_to_18(profile: ConfigObj):
 
 	section = "documentFormatting"
 	key = "reportSpellingErrors"
+	newKey = "reportSpellingErrors2"
 	try:
 		oldValue: bool = profile[section].as_bool(key)
 	except KeyError:
 		log.debug(f"'{key}' not present in config, no action taken.")
 		return
 	except ValueError:
-		log.error(f"'{key}' is not a boolean, got {profile[section][key]!r}. Deleting.")
-		del profile[section][key]
+		log.error(f"'{key}' is not a boolean, got {profile[section][key]!r}. No action taken.")
 		return
-	else:
-		newValue = ReportSpellingErrors.SPEECH.value if oldValue else ReportSpellingErrors.OFF.value
-		profile[section][key] = newValue
-		log.debug(
-			f"Converted '{key}' from {oldValue!r} to {newValue} ({ReportSpellingErrors(newValue).name}).",
-		)
+
+	del profile[section][key]
+	newValue = ReportSpellingErrors.SPEECH.value if oldValue else ReportSpellingErrors.OFF.value
+	profile[section][newKey] = newValue
+	log.debug(
+		f"Converted '{key}' from {oldValue!r} to {newValue} ({ReportSpellingErrors(newValue).name}).",
+	)