+++ source/speech/speech.py
@@ -1516,7 +1516,7 @@ def getTextInfoSpeech(  # noqa: C901
 	)
 	# For performance reasons, when navigating by paragraph or table cell, spelling errors will not be announced.
 	if unit in (textInfos.UNIT_PARAGRAPH, textInfos.UNIT_CELL) and reason == OutputReason.CARET:
-		formatConfig["reportSpellingErrors"] = ReportSpellingErrors.OFF.value
+		formatConfig["reportSpellingErrors2"] = ReportSpellingErrors.OFF.value
 
 	# Fetch the last controlFieldStack, or make a blank one
 	controlFieldStackCache = speakTextInfoState.controlFieldStackCache if speakTextInfoState else []
@@ -2993,12 +2993,12 @@ def getFormatFieldSpeech(  # noqa: C901
 				# Translators: Reported when text no longer contains a bookmark
 				text = _("out of bookmark")
 				textList.append(text)
-	if formatConfig["reportSpellingErrors"] != ReportSpellingErrors.OFF.value:
+	if formatConfig["reportSpellingErrors2"] != ReportSpellingErrors.OFF.value:
 		invalidSpelling = attrs.get("invalid-spelling")
 		oldInvalidSpelling = attrsCache.get("invalid-spelling") if attrsCache is not None else None
 		if (invalidSpelling or oldInvalidSpelling is not None) and invalidSpelling != oldInvalidSpelling:
 			if invalidSpelling:
-				if formatConfig["reportSpellingErrors"] == ReportSpellingErrors.SPEECH.value:
+				if formatConfig["reportSpellingErrors2"] == ReportSpellingErrors.SPEECH.value:
 					# Translators: Reported when text contains a spelling error.
 					text = _("spelling error")
 				else: